Moving and Stationary Object Detection using Depth
Background subtraction is a technique for separating out the foreground elements from the background. Background subtraction is a major preprocessing step for moving object detection in many vision based applications such as video surveillance systems, human-computer interaction, optical motion capture, etc. Background subtraction segments moving objects by obtaining the difference between the background image and the current image. It is mainly used in many computer vision applications as a primary step for object recognition, classification, and activity analysis tasks. Background subtraction can be applied for extracting information about the particular vehicle with the help of images provided by traffic cameras as well as content-based video coding, real-time motion gesture recognition or to keep a track of person, and further, it can be used in autonomous vehicles, etc. Here as part of our course project we have investigated some of the key points for the foreground/background separation when the depth of the image is also provided. The dataset chosen to carry out the investigation is the SBM-RGBD dataset as it contains inputs, ground-truth(GT) and depth. It sizes 1.11 GB.